pyproject.toml
[project]
name = "creditcoin-payout"
version = "1.0.0"
description = "Creditcoin Validator 자동 Payout 프로그램"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "substrate-interface>=1.7.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0",
    "structlog>=24.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
]

[project.scripts]
creditcoin-payout = "creditcoin_payout.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
